Update README.md; Small code change for compatibility; Python copyLib.py change for compatibility; Update CMake

+++ interface/python/copyLib.py
@@ -13,7 +13,9 @@ def init_module(dll_path='../../cmake-build-debug/libVParserPy.dylib'):
     """
     absPath = os.path.abspath('.')
     dlibPath = os.path.join(absPath, dll_path)
-    shutil.copy(dlibPath, os.path.join(absPath, 'LapisParser.so'))
+    if os.path.exists(dlibPath):
+        shutil.copy(dlibPath, os.path.join(absPath, 'LapisParser.so'))
 
 if __name__ == '__main__':
     init_module()
+    init_module('../../libVParserPy.so')
